fixes calendar abbreviations to be the campaign locale instead of the browsers locale.
I am getting the locale from PHP and setting it as an attribute on the components. Because DT has an underscore in the locale names instead of a dash I am replacing that in the connectedCallbacks of the components. This could be refactored with a helper function in the future.